pyproject.toml
[tool.poetry]
name = "authup"
version = "0.5.1"
description = "Python plugins for the Authup authentication and authorization framework"
authors = ["Michael Graf <michael.graf3110@gmail.com>"]
license = "MIT"
repository = "https://github.com/migraf/authup-py.git"
readme = "README.md"
homepage = "https://authup.org"


[tool.poetry.dependencies]
python = "^3.7"
httpx = "*"
pydantic = "*"
fastapi = { version=">= 0.87.0", optional=true}
requests = {version="*", optional=true}
flask = {version="^2.0.0", optional=true}
asgiref = "^3.6.0"


[tool.poetry.extras]
fastapi = ["fastapi"]
requests = ["requests"]
flask = ["flask"]
asgi = ["starlette"]
all = ["fastapi", "requests", "flask"]


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.227"
pytest = "^7.2.0"
black = "^22.12.0"
coverage = "^7.0.1"
python-dotenv = "^0.21.0"
pytest-dotenv = "^0.5.2"
pytest-asyncio = "^0.20.3"
pre-commit = "^2.21.0"
tox = "*"


[tool.ruff]
line-length = 120


# Enable Pyflakes `E` and `F` codes by default.
select = ["E", "F", "I001"]
ignore = []
per-file-ignores = { "*/__init__.py" = ["F401"] }

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.10.
target-version = "py311"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.tox]
legacy_tox_ini = """
[tox]
envlist = py37, py38, py39, py310

[gh-actions]
python =
    3.7: py37
    3.8: py38
    3.9: py39
    3.10: py310

[testenv]
passenv = *
allowlist_externals = poetry
commands =
    poetry run pytest
"""

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
